+++ api/test/Helper.cs
@@ -5,29 +5,27 @@ using NUnit.Framework;
 
 namespace test;
 
-public static class Helper
+public class Helper
 {
     public static readonly Uri Uri;
-    public static readonly string ProperlyFormattedConnectionString;
+    private static readonly string ProperlyFormattedConnectionString;
     public static readonly NpgsqlDataSource DataSource;
 
     static Helper()
     {
         string rawConnectionString;
-        string envVarKeyName = "pgconn";
+        String envVarKeyName = "pgconn";
 
         rawConnectionString = Environment.GetEnvironmentVariable(envVarKeyName)!;
         if (rawConnectionString == null)
         {
-            throw new Exception($@"Connection string empty
-");
+            throw new Exception($@"The Connection String is empty.");
         }
 
         try
         {
             Uri = new Uri(rawConnectionString);
-            ProperlyFormattedConnectionString = string.Format(
-                "Server={0};Database={1};User Id={2};Password={3};Port={4};Pooling=true;MaxPoolSize=3",
+            ProperlyFormattedConnectionString = string.Format("Server={0};Database={1};User Id={2};Password={3};Port={4};Pooling=true;MaxPoolSize=3",
                 Uri.Host,
                 Uri.AbsolutePath.Trim('/'),
                 Uri.UserInfo.Split(':')[0],
@@ -39,15 +37,17 @@ public static class Helper
         }
         catch ( Exception e)
         {
-            throw new Exception($@"Connection string is there but something else went wrong.", e);
+            throw new Exception($@"Connection string found but there was an issue. The input may be incorrect.");
         }
     }
-    
+
     public static string BadResponseBody(string content)
     {
-        return $@"
-RESPONSE BODY: {content}
-";
+        return $@"There was an issue fetching the response body from the API and turning it into a class object.
+        Reponse Body: {content}
+
+        EXCEPTION:
+        ";
     }
 
     public static void TriggerRebuild()
@@ -60,36 +60,31 @@ RESPONSE BODY: {content}
             }
             catch (Exception e)
             {
-                throw new Exception($@"Something is wrong with the rebuild script.", e);
+                throw new Exception($@"
+                There was an error rebuilding the DB.
+                EXCEPTION: 
+                ", e);
             }
         }
-    }
-
-    public static string MyBecause(object actual, object expected)
-    {
-        string expectedJson = JsonConvert.SerializeObject(expected, Formatting.Indented);
-        string actualJson = JsonConvert.SerializeObject(actual, Formatting.Indented);
-
-        return $"because we want these objects to be equivalent:\nExpected:\n{expectedJson}\nActual:\n{actualJson}";
     }
 
     public static string RebuildScript = @"
 DROP SCHEMA IF EXISTS box_factory CASCADE;
 CREATE SCHEMA box_factory;
+
 create table if not exists box_factory.boxes
 (
-    id       integer generated by default as identity
-        constraint boxespk
-            primary key,
+    id       integer generated by default as identity,
     size     text,
     weight   integer,
     price    integer,
     material text,
     color    text,
-    quantity integer
-);
- ";
+    quantity integer,
+    constraint boxespk primary key (id)
+);";
 
-    public static string NoResponseMessage = $@"No answer from API, check connection.
-";
+    public static string NoResponseMessage = $@"
+There was no response from the API, the API may not be running.
+    ";
 }
