+++ api/test/Models.cs
@@ -10,3 +10,12 @@ public class Box
     public string Color { get; set; }
     public int Quantity { get; set; }
 }
+
+public class InStockBoxes
+{
+    public int Id { get; set; }
+    
+    public int Weight { get; set; }
+    
+    public int Quantity { get; set; }
+}
